+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
+## [1.4.0] - 2025-0-26
+### Added
+- Added Logging and Custom Logging
+- Shifted Auth requirements to the developer
+
 ## [1.3.1] - 2025-09-20
 ### Added
 - Signature Verification for auth